refactor(server): split admin runner structure
Separate server handlers and runner lifecycle code into focused files, group store interfaces by responsibility, and extract admin UI types and login page.
